requirements.txt
diff-match-patch==20121119
Django==2.0.6
django-import-export==1.0.1
djangorestframework==3.8.2
Pillow==5.1.0
psycopg2-binary